requirements.txt
langgraph~=0.2.59
python-dotenv~=1.0.1
langchain-community~=0.3.12
langchain-core~=0.3.25
rich~=13.9.4
Telethon~=1.38.1
typing_extensions~=4.12.2
langchain-anthropic~=0.3.0
python-telegram-bot~=21.9
numpy~=1.26.4
torch~=2.5.1
sentence-transformers~=3.3.1
scikit-learn~=1.6.0
transformers~=4.47.1
chromadb~=0.5.20
requests~=2.32.3
pymongo~=4.10.1
groq~=0.13.1
langchain-experimental~=0.3.3